requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
black==24.8.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.1
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.7
eventlet==0.37.0
flake8==7.1.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13